+++ Sources/GitHubAccountSwitcherCore/AppLinks.swift
@@ -10,4 +10,6 @@ public enum AppLinks {
     public static let newIssue = URL(
         string: "https://github.com/mobilepur/github-account-switcher/issues/new"
     )!
+
+    public static let deviceSignIn = URL(string: "https://github.com/login/device")!
 }